DEMOS-2117: added insert to application tags
from uipath value
